Organized files in the repository into two seperate folders, "Scripts" and "Profile". Changed "domainJoin.ps1" to allow the use of any domain to be entered by the script user.
The "Profile" folder will now contain the powershell profile so it is not
mixed with the scripts.

The "Scripts folder will contain all of the custom Powershell scripts.

domainJoin.ps1 was modified to prompt the script user to enter the desired
domain name into the program, rather than having it hardcoded into the
script.